% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-flea.R
\docType{data}
\name{flea}
\alias{flea}
\title{Flea Beatles Measurements}
\format{
A data frame with 74 observations and 7 variables:
\describe{
\item{tars1}{width of the first joint of the first tarsus in microns (the sum of measurements for both tarsi)}
\item{tars2}{ the same for the second joint}
\item{head}{the maximal width of the head between the external edges of the eyes in 0.01 mm}
\item{aede1}{the maximal width of the aedeagus in the fore-part in microns}
\item{aede2}{the front angle of the aedeagus ( 1 unit = 7.5 degrees)}
\item{aede3}{the aedeagus width from the side in microns}
\item{species}{which species is being examined - concinna, heptapotamica, heikertingeri}
}
}
\usage{
flea
}
\description{
This data is from a paper by A. A. Lubischew, "On the Use of Discriminant Functions in Taxonomy",
Biometrics, Dec 1962, pp.455-477.
}
\examples{
head(flea)

}
\keyword{datasets}
